Повышенная сложность

+++ src/Employee.java
@@ -1,12 +1,12 @@
 public class Employee {
     private String name;
-    private int salary;
+    private double salary;
     private int department;
     private int id;
     public static int idCounter = 1;
 
 
-    public Employee(String name, int department, int salary) {
+    public Employee(String name, int department, double salary) {
         this.name = name;
         this.salary = salary;
         this.department = department;
@@ -18,7 +18,7 @@ public class Employee {
         return name;
     }
 
-    public int getSalary() {
+    public double getSalary() {
         return salary;
     }
 
@@ -30,7 +30,7 @@ public class Employee {
         return id;
     }
 
-    public void setSalary(int salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
